Extract Signal interface for simple computed signals
Fixes #22121
