+++ flow-server/src/test/java/com/vaadin/flow/server/VaadinServiceTest.java
@@ -74,8 +74,8 @@ import com.vaadin.flow.server.communication.WebComponentBootstrapHandler;
 import com.vaadin.flow.server.communication.WebComponentProvider;
 import com.vaadin.flow.server.menu.AvailableViewInfo;
 import com.vaadin.flow.server.startup.ApplicationRouteRegistry;
+import com.vaadin.signals.AbstractSignal;
 import com.vaadin.signals.ListSignal;
-import com.vaadin.signals.Signal;
 import com.vaadin.signals.SignalEnvironment;
 import com.vaadin.tests.util.AlwaysLockedVaadinSession;
 import com.vaadin.tests.util.MockDeploymentConfiguration;
