+++ signals/src/main/java/com/vaadin/signals/impl/ComputedSignal.java
@@ -21,6 +21,7 @@ import java.util.function.Supplier;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.POJONode;
+import com.vaadin.signals.AbstractSignal;
 import com.vaadin.signals.Id;
 import com.vaadin.signals.Node.Data;
 import com.vaadin.signals.NodeSignal;
@@ -39,7 +40,7 @@ import com.vaadin.signals.impl.UsageTracker.Usage;
  * @param <T>
  *            the value type
  */
-public class ComputedSignal<T> extends Signal<T> {
+public class ComputedSignal<T> extends AbstractSignal<T> {
 
     /*
      * This state is never supposed to be synchronized across a cluster or to
