+++ signals/src/test/java/com/vaadin/signals/impl/ComputedSignalTest.java
@@ -29,6 +29,7 @@ import java.util.concurrent.locks.LockSupport;
 
 import org.junit.jupiter.api.Test;
 
+import com.vaadin.signals.AbstractSignal;
 import com.vaadin.signals.Signal;
 import com.vaadin.signals.SignalTestBase;
 import com.vaadin.signals.ValueSignal;
@@ -268,7 +269,8 @@ public class ComputedSignalTest extends SignalTestBase {
 
     @Test
     void unsuppotedOperations_runOperations_throws() {
-        Signal<Object> signal = Signal.computed(() -> null);
+        AbstractSignal<Object> signal = (AbstractSignal<Object>) Signal
+                .computed(() -> null);
 
         assertThrows(UnsupportedOperationException.class, () -> {
             signal.peek();
@@ -277,6 +279,30 @@ public class ComputedSignalTest extends SignalTestBase {
         assertThrows(UnsupportedOperationException.class, () -> {
             signal.peekConfirmed();
         });
+    }
+
+    @Test
+    void lambda_computesValue_computedNotCached() {
+        ValueSignal<Integer> signal = new ValueSignal<>(1);
+
+        AtomicInteger count = new AtomicInteger();
+
+        Signal<Integer> doubled = () -> {
+            count.incrementAndGet();
+            return signal.value() * 2;
+        };
+
+        assertEquals(2, doubled.value());
+        assertEquals(1, count.intValue());
+
+        assertEquals(2, doubled.value());
+        assertEquals(2, count.intValue());
+
+        signal.value(3);
+        assertEquals(2, count.intValue());
+
+        assertEquals(6, doubled.value());
+        assertEquals(3, count.intValue());
     }
 
     private static boolean waitForGarbageCollection(WeakReference<?> ref) {
